GCP-App-Engine-standard/Python-bigquery/requirements.txt
google-cloud-bigquery==2.34.4
Flask==2.1.0
pandas
pyarrow